+++ shared/consts.hpp
@@ -2,4 +2,4 @@
 
 static const int ChN = 2;
 static const int SampleRate = 44100;
-static const float DefaultBpm = 120;
+static const double DefaultBpm = 120;
